Narrow mesh control shift field to u8
Standardize `shift` as `u8` across the execute-controller/mesh control path (`ExCtrlMeshCntlPack`, queue payload, shared port structs, and testbench driver). This removes the unnecessary 32-bit cast in the packer and keeps the interface types consistent end-to-end.

+++ smesh/include/ExCtrlMeshCntlPack.hpp
@@ -57,7 +57,7 @@ class ExCtrlMeshCntlPack : public Component {
   Input(SmeshRsTag, rs_tag);
   Input(u8, dataflow);
   Input(bit, prop);
-  Input(u32, shift);
+  Input(u8, shift);
   Input(bit, im2colling);
   Input(bit, first);
